requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
streamlit==1.28.0

# LangChain - Pinned compatible versions for Python 3.12
langchain==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0

# Utilities
python-dotenv==1.0.0
pydantic==2.9.0
requests==2.31.0
aiofiles==23.2.1

